test: split admin marketplace tests
